requirement.txt
google-generativeai
opencv-python
pyttsx3
speechrecognition
pyaudio
python-dotenv